[backend/frontend] refactor(Subscription): improve ServiceSlug (#2159)
